--- teca/src/app/shared/components/notifications-bar/notification-bar.component.html
@@ -1,12 +1,0 @@
-<ion-grid class="box "   >
-  
-
-<ion-col  >
-  <button style="background-color: #ffffff !important"id="notifications-outline-btn" >
-    <ion-icon name="notifications-outline"></ion-icon>
-    <ion-badge color = "red "id="notifications-outline">1</ion-badge><!-- cantidad de notificaciones -->
-</button>
-</ion-col>
-
-
-</ion-grid>
