--- teca/src/app/shared/components/notifications-bar/notification-bar.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-NotificationBar',
-  templateUrl: './notification-bar.component.html',
-  styleUrls: ['./notification-bar.component.scss'],
-})
-export class NotificationBar implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {}
-
-}
