+++ teca/src/app/shared/components/profile-bar/notifications-bar/notification-bar.component.html
@@ -1,0 +1,15 @@
+<ion-grid class="box">
+  <ion-col>
+    <button
+      style="background-color: #ffffff !important"
+      id="notifications-outline-btn"
+    >
+      <ion-icon name="notifications-outline"></ion-icon>
+      <ion-badge color="red " id="notifications-outline">1</ion-badge
+      ><!-- cantidad de notificaciones -->
+    </button>
+    <ion-label class="name">{{ currentUser.name }}</ion-label>
+
+    <ion-img [src]="currentUser.avatar" class="avatar"></ion-img>
+  </ion-col>
+</ion-grid>
